% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.sketches-package.R
\docType{package}
\name{data.sketches-package}
\alias{data.sketches}
\alias{data.sketches-package}
\title{data.sketches: Probabilistic Streaming Data Sketches}
\description{
Provides an interface to the 'Apache DataSketches' (\url{https://datasketches.apache.org/}) library of streaming algorithms for approximate analytics on data too large to hold or process exactly. Sketches are compact, mergeable summaries built in a single pass over a stream that answer queries such as approximate distinct counts, quantiles and ranks, frequent items and point-frequency estimates, weighted sampling, and set membership with mathematically proven error bounds. Implements Karnin-Lang-Liberty (KLL), Relative Error Quantiles (REQ), t-Digest, HyperLogLog (HLL), Compressed Probabilistic Counting (CPC), Theta, Frequent Items, Count-Min, Array of Doubles, Variance Optimal (VarOpt), Exact and Bounded Probabilistic Proportional-to-Size (EBPPS), and Bloom filter sketches, with native serialization for interoperability with other 'Apache DataSketches' implementations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pedrobtz/data.sketches}
  \item \url{https://pedrobtz.github.io/data.sketches/}
  \item Report bugs at \url{https://github.com/pedrobtz/data.sketches/issues}
}

}
\author{
\strong{Maintainer}: Pedro Baltazar \email{pedrobtz@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item The Apache Software Foundation (Author and copyright holder of bundled Apache DataSketches C++ code) [contributor, copyright holder]
  \item Yahoo Inc. (Copyright holder of portions of bundled Apache DataSketches C++ code) [copyright holder]
  \item Verizon Media (Copyright holder of portions of bundled Apache DataSketches C++ code) [copyright holder]
  \item Stephan Brumme (Author and copyright holder of bundled xxhash64.h code) [contributor, copyright holder]
  \item Austin Appleby (Author of bundled public-domain MurmurHash3 code) [contributor]
  \item Sean Eron Anderson (Author of bundled public-domain bit-hack code used in ceiling_power_of_2.hpp) [contributor]
}

}
\keyword{internal}
